lot of bug
the name of the columns address mail is in french so I needed to change it.
The dataframe need to be in float otherwise it's a mess.
We do not ask people to put the same address mail than in the registration so I introduce the check of the mail address as optionnal